Try to fix transform anim duration seconds

+++ engine/core/include/animations/bone_animation.h
@@ -56,6 +56,6 @@ namespace engine
 		void readMissingBones(const aiAnimation* animation, Model& model);
 		void readHierarchyData(AssimpNodeData& dest, const aiNode* src);
 
-		double computeFPS(const aiAnimation* anim);
+		unsigned int computeFPS(const aiAnimation* anim);
 	};
 }
